Fix display of last report when it was today
